refactor: 메세지, 여행, 위시리스트 앱라우트 정리
