+++ src/app/(guest)/components/reservation/PaymentInformation.stories.tsx
