+++ src/app/(guest)/components/reservation/ReservationContent.tsx
@@ -1,9 +1,9 @@
-import { CustomerSupport } from '@/app/messages/components/reservation/CustomerSupport';
-import { HostInformation } from '@/app/messages/components/reservation/HostInformation';
-import { PaymentInformation } from '@/app/messages/components/reservation/PaymentInformation';
-import { ReservationDetailInformation } from '@/app/messages/components/reservation/ReservationDetailInformation';
-import { RoomInformation } from '@/app/messages/components/reservation/RoomInformation';
-import { RoomRules } from '@/app/messages/components/reservation/RoomRules';
+import { CustomerSupport } from '@/app/(guest)/components/reservation/CustomerSupport';
+import { HostInformation } from '@/app/(guest)/components/reservation/HostInformation';
+import { PaymentInformation } from '@/app/(guest)/components/reservation/PaymentInformation';
+import { ReservationDetailInformation } from '@/app/(guest)/components/reservation/ReservationDetailInformation';
+import { RoomInformation } from '@/app/(guest)/components/reservation/RoomInformation';
+import { RoomRules } from '@/app/(guest)/components/reservation/RoomRules';
 
 export function ReservationContent() {
   return (
