+++ src/app/(guest)/components/reservation/RoomInformation.stories.tsx
