+++ src/app/(guest)/components/reservation/RoomInformation.tsx
