+++ src/app/(guest)/trips/page.tsx
@@ -1,0 +1,3 @@
+export default function Trips() {
+  return <h1>여행 페이지</h1>
+}
