+++ src/app/(guest)/wishlists/page.tsx
@@ -1,0 +1,3 @@
+export default function Wishlists() {
+  return <h1>위시리스트 페이지</h1>
+}
